+++ src/js/helpers/fire-ga-event.js
@@ -1,14 +1,18 @@
 module.exports = function( eventCategory, eventAction ) {
   var TRACKER_NAME = 'gtm1';
-  var getRPName = require( 'helpers/getRPName' );
+  var getRPName = require( 'helpers/get-rp-name' );
   var RP = getRPName();
 
-  if ( window.ga ) {
+  if ( window.ga && NLX.environment === 'PROD' ) {
     ga( TRACKER_NAME + '.send', {
       'hitType': 'event',
       'eventCategory': eventCategory,
       'eventAction': eventAction,
       'eventLabel': RP
     });
+  }
+
+  if ( NLX.environment !== 'PROD' ) {
+    console.log( eventCategory + ' - ' + eventAction );
   }
 };