+++ src/js/helpers/store-last-used-connection.js
@@ -1,0 +1,9 @@
+var accountLinking = require( 'helpers//account-linking' );
+
+module.exports = function( connectionName ) {
+  var isAccountLinking = accountLinking.isAccountLinking();
+
+  if ( !isAccountLinking && window.localStorage ) {
+    window.localStorage.setItem( 'nlx-last-used-connection', connectionName );
+  }
+};