+++ src/js/main.js
@@ -11,6 +11,7 @@ document.documentElement.className = 'has-js';
 
 // environment variables are replaced in build process
 window.NLX = {
+  'environment': '{{{ environment }}}',
   'auth0_domain': '{{{ auth0_domain }}}',
   'client_ID': '{{{ client_ID }}}',
   'LDAP_connection_name': '{{{ LDAP_connection_name }}}',
@@ -19,7 +20,8 @@ window.NLX = {
   'features': {
     'autologin': '{{{ features.autologin }}}',
     'person_api_lookup': '{{{ features.person_api_lookup }}}'
-  }
+  },
+  'supportedLoginMethods': [ 'github', 'google-oauth2', 'email' ]
 };
 
 polyfill();